+++ wm-system/product/src/main/kotlin/com/ajaxproject/warehouse/application/port/api/CreateProductApiInPort.kt
@@ -1,4 +1,4 @@
-package com.ajaxproject.warehouse.application.port
+package com.ajaxproject.warehouse.application.port.api
 
 import com.ajaxproject.api.internal.warehousesvc.input.reqreply.product.CreateProductRequest
 import com.ajaxproject.api.internal.warehousesvc.input.reqreply.product.CreateProductResponse
