+++ wm-system/product/src/main/kotlin/com/ajaxproject/warehouse/infrastructure/adapter/common/service/GetProductByIdApiService.kt
@@ -1,7 +1,7 @@
 package com.ajaxproject.warehouse.infrastructure.adapter.common.service
 
 import com.ajaxproject.api.internal.warehousesvc.input.reqreply.product.GetProductByIdResponse
-import com.ajaxproject.warehouse.application.port.GetProductByIdApiInPort
+import com.ajaxproject.warehouse.application.port.api.GetProductByIdApiInPort
 import com.ajaxproject.warehouse.application.port.ProductServiceInPort
 import com.ajaxproject.warehouse.domain.Product
 import com.ajaxproject.warehouse.infrastructure.mapper.mapToProto
