+++ wm-system/product/src/main/kotlin/com/ajaxproject/warehouse/infrastructure/adapter/nats/controller/GetProductByIdNatsController.kt
@@ -3,7 +3,7 @@ package com.ajaxproject.warehouse.infrastructure.adapter.nats.controller
 import com.ajaxproject.api.internal.warehousesvc.NatsSubject
 import com.ajaxproject.api.internal.warehousesvc.input.reqreply.product.GetProductByIdRequest
 import com.ajaxproject.api.internal.warehousesvc.input.reqreply.product.GetProductByIdResponse
-import com.ajaxproject.warehouse.application.port.GetProductByIdApiInPort
+import com.ajaxproject.warehouse.application.port.api.GetProductByIdApiInPort
 import com.ajaxproject.warehouse.nats.NatsController
 import com.google.protobuf.Parser
 import org.springframework.stereotype.Component
